requirements_dev.txt
PyYAML==3.12
Sphinx==1.7.0b2
boto3==1.5.24
bumpversion==0.5.3
coverage==4.5
django==2.0.2
fabric3==1.13.1.post1
flake8==3.5.0
pelican==3.7.1
python-dotenv==0.7.1
recommonmark==0.4.0  # To support Markdown in docs
tox==3.0.0rc1
unipath==1.1
watchdog==0.8.3
wheel==0.30.0
